pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "moviepy>=2.2.1",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "streamlit>=1.49.1",
]
